refactor: modularize code and fix maze generation bug
