+++ src/position.rs
@@ -1,0 +1,12 @@
+// 位置情報を表す構造体
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct Position {
+    pub x: usize,
+    pub y: usize,
+}
+
+impl Position {
+    pub fn new(x: usize, y: usize) -> Self {
+        Position { x, y }
+    }
+}
